+++ theme/js/components/discussions/threads-create.vue
@@ -89,9 +89,7 @@ export default {
   },
   methods: {
     displayForm() {
-      this.$auth(
-        this.$t("You must be logged in to start a discussion.")
-      );
+      this.$auth();
       this.showForm = true;
     },
     hideForm() {