+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 1.2.2 (2022-01-21)
 
 - Fix latest modification date for dataset and resources on dataset page [#62](https://github.com/etalab/udata-front/pull/62)